Corrected wrong behaviour on FAQ section with show more button
The issue was caused because the resize() function was containing a script that was hiding the desired elements, but resize() function was being called everytime the user was opening or closing a faq element, and everytime the user's screens was resizing. Also, an extra button with "show less" text has been added and now the correct button shows/hides on appropiate events.